requirements.txt
asgiref==3.5.2
Django==4.0.6
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
